@@ -3,7 +3,7 @@
     "version": "0.0.0",
     "scripts": {
         "ng": "ng",
-        "start": "ng serve",
+        "start": "ng serve --poll 2000",
         "build": "ng build",
         "build:prod": "ng build --prod",
         "build:staging": "ng build -c staging",
@@ -41,22 +41,35 @@
         "@ngrx/store-devtools": "^6.1.1",
         "@ngx-translate/core": "^11.0.1",
         "@ngx-translate/http-loader": "^4.0.0",
-        "@swimlane/ngx-charts": "8.1.0",
+        "@swimlane/ngx-charts": "8.0.0",
         "@swimlane/ngx-graph": "5.0.0",
         "@uppy/core": "^0.29.1",
         "@uppy/drag-drop": "^0.29.1",
         "animate.css": "^3.7.0",
+        "babel-plugin-idx": "^2.4.0",
         "bootstrap": "^4.1.3",
+        "brace": "^0.11.1",
         "core-js": "^2.5.4",
         "firebase": "^5.8.3",
         "hammerjs": "^2.0.8",
+        "highlightjs": "^9.12.0",
+        "idx": "^2.5.4",
         "javascript-time-ago": "^2.0.1",
         "moment": "^2.24.0",
         "ngx-bootstrap": "^3.1.3",
         "ngx-gravatar": "^7.2.0",
+        "ngx-markdown": "^7.1.4",
+        "ngx-markdown-editor": "^1.2.0",
         "ngx-mask": "^7.3.5",
+        "ngx-monaco-editor": "^7.0.0",
+        "ngx-quill": "^4.6.12",
         "ngx-simple-modal": "^1.3.13",
         "ngx-toastr": "^9.1.1",
+        "ngx-youtube-player": "^6.0.0",
+        "prism-theme-one-dark": "^1.0.0",
+        "prismjs": "^1.15.0",
+        "qrcode": "^1.3.3",
+        "quill": "^1.3.6",
         "rxjs": "~6.3.3",
         "socket.io-client": "^2.1.1",
         "uuid": "^3.3.2",
@@ -72,6 +85,7 @@
         "@types/jasmine": "~2.8.8",
         "@types/jasminewd2": "~2.0.3",
         "@types/node": "~8.9.4",
+        "@types/qrcode": "^1.3.2",
         "@types/socket.io-client": "^1.4.32",
         "codelyzer": "~4.5.0",
         "compression": "^1.7.3",
@@ -89,6 +103,6 @@
         "protractor": "~5.4.0",
         "ts-node": "~7.0.0",
         "tslint": "~5.11.0",
-        "typescript": "~3.1.1"
+        "typescript": "~3.2.4"
     }
 }
