@@ -1,8 +1,8 @@
 import { Component, Input, OnInit } from "@angular/core";
 import { Store, select } from "@ngrx/store";
 import { Observable } from "rxjs";
-import { map } from "rxjs/operators";
-import { getPuzzleHeroInfo, State } from "src/app/store/app.reducers";
+import { map, withLatestFrom } from "rxjs/operators";
+import { getPuzzleHeroInfo, State, getRegisteredCompetitions } from "src/app/store/app.reducers";
 import * as fromApp from "src/app/store/app.reducers";
 import { Router } from "@angular/router";
 
@@ -16,16 +16,29 @@ export class SideNavComponent implements OnInit {
     attendee$ = this.store$.pipe(select(fromApp.getCurrentAttendee));
     currentEvent$ = this.store$.pipe(select(fromApp.getCurrentEvent));
     puzzleHeroInfo$ = this.store$.pipe(select(getPuzzleHeroInfo));
+    registeredCompetitions$ = this.store$.pipe(select(getRegisteredCompetitions));
 
     get puzzleHeroOpen$(): Observable<boolean> {
         return this.puzzleHeroInfo$.pipe(
-            map(info => info.open)
+            withLatestFrom(this.attendee$),
+            map(([info, attendee]) => info && info.open || attendee && (attendee.role === "admin" || attendee.role === "super-admin"))
         );
     }
 
     get scoreboardOpen$(): Observable<boolean> {
         return this.puzzleHeroInfo$.pipe(
-            map(info => info.scoreboardOpen)
+            withLatestFrom(this.attendee$),
+            map(([info, attendee]) => info && info.scoreboardOpen || attendee &&
+                (attendee.role === "admin" || attendee.role === "super-admin"))
+        );
+    }
+
+    get flashoutOpen$(): Observable<boolean> {
+        return this.currentEvent$.pipe(
+            map(e => {
+                const now = new Date().toISOString();
+                return now > e.flashoutBeginDate && now < e.flashoutEndDate;
+            })
         );
     }
 
@@ -34,7 +47,7 @@ export class SideNavComponent implements OnInit {
 
     constructor(private store$: Store<State>, private router: Router) { }
 
-    ngOnInit() {}
+    ngOnInit() { }
 
     isActive(route: string) {
         return this.router.isActive(route, false);
