@@ -3,20 +3,24 @@ import { CanActivate, Router } from "@angular/router";
 import { select, Store } from "@ngrx/store";
 import { Observable, of } from "rxjs";
 import { fromPromise } from "rxjs/internal-compatibility";
-import { filter, switchMap } from "rxjs/operators";
-import { getPuzzleHeroInfo, State } from "../../../store/app.reducers";
+import { filter, switchMap, withLatestFrom } from "rxjs/operators";
+import { getCurrentAttendee, getPuzzleHeroInfo, State } from "../../../store/app.reducers";
+import { PuzzleHeroInfo } from "../../../api/models/puzzle-hero";
+import { Attendee } from "../../../api/models/attendee";
 
 @Injectable()
 export class ScoreboardGuard implements CanActivate {
     private puzzleHeroInfo$ = this.store$.pipe(select(getPuzzleHeroInfo));
+    private currentAttendee$ = this.store$.pipe(select(getCurrentAttendee));
 
     constructor(private store$: Store<State>, private router: Router) {}
 
     public canActivate(): Observable<boolean> {
         return this.puzzleHeroInfo$.pipe(
             filter(x => x !== null),
-            switchMap(info => {
-                if (info.scoreboardOpen) {
+            withLatestFrom(this.currentAttendee$),
+            switchMap(([info, attendee]: [PuzzleHeroInfo, Attendee]) => {
+                if (info.scoreboardOpen || attendee && attendee.role === "admin") {
                     return of(true);
                 } else {
                     return fromPromise(this.router.navigate(["/puzzle-hero/tracks"]));
