@@ -12,6 +12,8 @@ import { UserApi } from "./user.api";
 import { SchoolApi } from "./school.api";
 import { PuzzleHeroApi } from "./puzzle-hero.api";
 import { ActivityApi } from "./activity.api";
+import { FlashoutApi } from "./flashout.api";
+import { CompetitionApi } from "./competition.api";
 
 @NgModule({
     imports: [
@@ -29,7 +31,9 @@ import { ActivityApi } from "./activity.api";
         PuzzleHeroApi,
         ApiService,
         UserApi,
-        ActivityApi
+        ActivityApi,
+        CompetitionApi,
+        FlashoutApi
     ]
 })
 export class ApiModule {}
