@@ -1,20 +1,27 @@
 import { EventGuide } from "./guide";
+import { TeamResult } from "./competition";
 
 export interface Event {
-    _id: string;
-    name: string;
+    _id?: string;
+    name?: string;
     details?: any;
-    beginDate: Date | string;
-    endDate: Date | string;
+    beginDate?: Date | string;
+    endDate?: Date | string;
+    flashoutBeginDate?: string | Date;
+    flashoutEndDate?: string | Date;
     activities?: any[];
-    imageUrl: string;
+    imageUrl?: string;
     coverUrl?: string;
     website?: string;
-    facebookEvent?: string;
     locationName?: string;
     locationAddress?: string;
     maxTeamMembers?: number;
-    guide: EventGuide;
-    teamEditLocked: boolean;
-    teamEditLockDate: Date | string;
+    guide?: EventGuide;
+    teamEditLocked?: boolean;
+    teamEditLockDate?: Date | string;
+}
+
+export interface EventScore {
+    overall: TeamResult[];
+    competitions: { _id: string, name: any, results: TeamResult[] }[];
 }
