@@ -3,11 +3,9 @@ import { AngularFireMessagingModule } from "@angular/fire/messaging";
 import { BrowserModule } from "@angular/platform-browser";
 import { NgModule } from "@angular/core";
 import { ServiceWorkerModule } from "@angular/service-worker";
-
 import { AppComponent } from "./app.component";
 import { StoreModule } from "@ngrx/store";
 import { EffectsModule } from "@ngrx/effects";
-
 import * as fromApp from "./store/app.reducers";
 import { AppEffects } from "./store/app.effects";
 import { RouterModule } from "@angular/router";
@@ -32,11 +30,21 @@ import { TeamService } from "./providers/team.service";
 import { EventService } from "./providers/event.service";
 import { RegisterService } from "./providers/register.service";
 import { ForgetModule } from "./features/forget/forget.module";
+import { MarkdownModule } from "ngx-markdown";
 import { PasswordService } from "./providers/password.service";
 import { ToastrModule } from "ngx-toastr";
 import { NgxMaskModule } from "ngx-mask";
 import { ResetModule } from "./features/reset/reset.module";
-import { AccordionModule, TooltipModule } from "ngx-bootstrap";
+import { MonacoEditorModule } from "ngx-monaco-editor";
+import {
+    AccordionModule,
+    PopoverModule,
+    AlertModule,
+    TooltipModule,
+    TimepickerModule,
+    DatepickerModule,
+    BsDatepickerModule, SortableModule, TabsModule
+} from "ngx-bootstrap";
 import { SponsorsService } from "./providers/sponsors.service";
 import { ScheduleService } from "./providers/schedule.service";
 import { RoleGuard } from "./guards/role.guard";
@@ -44,6 +52,11 @@ import { SchoolService } from "./providers/school.service";
 import { PuzzleHeroService } from "./providers/puzzle-hero.service";
 import { NotificationService } from "./providers/notification.service";
 import { UpdateModule } from "./update/update.module";
+import { FlashoutService } from "./providers/flashout.service";
+import "brace";
+import "brace/mode/markdown";
+import "brace/theme/github";
+import { CompetitionsService } from "./providers/competitions.service";
 
 export function loadFactory(http: HttpClient): TranslateHttpLoader {
     return new TranslateHttpLoader(http, "../assets/i18n/", ".json");
@@ -58,6 +71,8 @@ export function loadFactory(http: HttpClient): TranslateHttpLoader {
         BsDropdownModule.forRoot(),
         TooltipModule.forRoot(),
         AccordionModule.forRoot(),
+        PopoverModule.forRoot(),
+        AlertModule.forRoot(),
         DashboardModule,
         LoginModule,
         RegisterModule,
@@ -71,7 +86,13 @@ export function loadFactory(http: HttpClient): TranslateHttpLoader {
                 deps: [HttpClient]
             }
         }),
+        BsDatepickerModule.forRoot(),
+        DatepickerModule.forRoot(),
+        TimepickerModule.forRoot(),
+        SortableModule.forRoot(),
         NgxMaskModule.forRoot(),
+        TabsModule.forRoot(),
+        MarkdownModule.forRoot(),
         ApiModule,
         ForgetModule,
         ResetModule,
@@ -84,6 +105,7 @@ export function loadFactory(http: HttpClient): TranslateHttpLoader {
         !environment.production ? StoreDevtoolsModule.instrument() : [],
         BrowserAnimationsModule,
         ToastrModule.forRoot(),
+        MonacoEditorModule.forRoot(),
         NoopAnimationsModule,
         UpdateModule
     ],
@@ -101,9 +123,10 @@ export function loadFactory(http: HttpClient): TranslateHttpLoader {
         SponsorsService,
         PuzzleHeroService,
         ScheduleService,
-        NotificationService
+        NotificationService,
+        FlashoutService,
+        CompetitionsService
     ],
     bootstrap: [AppComponent]
 })
-export class AppModule {
-}
+export class AppModule {}
