@@ -8,7 +8,9 @@ import {
 } from "./store/tracks.reducer";
 import { select, Store } from "@ngrx/store";
 import { LoadStarredTracks, LoadTracks, StarTrack } from "./store/tracks.actions";
-import { Track } from "../../../api/models/puzzle-hero";
+import { PuzzleInfo, Track } from "../../../api/models/puzzle-hero";
+import { InfoPuzzleHeroComponent } from "./components/info-puzzle-hero/info-puzzle-hero.component";
+import { SimpleModalService } from "ngx-simple-modal";
 
 @Component({
     selector: "app-puzzle-hero",
@@ -24,16 +26,21 @@ export class TracksComponent implements OnInit {
 
     openTracks: Track[] = [];
 
-    constructor(private store$: Store<State>) {
+    constructor(private store$: Store<State>, private modalService: SimpleModalService) {
         this.store$.dispatch(new LoadTracks());
         this.store$.dispatch(new LoadStarredTracks());
     }
 
-    ngOnInit() {
-    }
+    ngOnInit() {}
 
     clickStar(track: Track) {
         this.store$.dispatch(new StarTrack(track));
+    }
+
+    clickPuzzle(puzzle: PuzzleInfo, track: Track) {
+        if (!puzzle.locked) {
+            this.modalService.addModal(InfoPuzzleHeroComponent, { puzzle, track });
+        }
     }
 
     trackOpenChange(track: Track, open: boolean) {
