@@ -4,11 +4,8 @@ import { CommonModule } from "@angular/common";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { TracksRoutingModule } from "./tracks-routing.module";
 import { RouterModule } from "@angular/router";
-import { TrackComponent } from "./components/track/track.component";
 import { AccordionModule } from "ngx-bootstrap";
 import { NgxGraphModule } from "@swimlane/ngx-graph";
-import { TrackDirective } from "./components/track/track.directive";
-import { PuzzleTileComponent } from "./components/puzzle-tile/puzzle-tile.component";
 import { FlexLayoutModule } from "@angular/flex-layout";
 import { StoreModule } from "@ngrx/store";
 import * as fromPuzzleHero from "./store/tracks.reducer";
@@ -17,6 +14,7 @@ import { TracksEffects } from "./store/tracks.effects";
 import { TranslateModule } from "@ngx-translate/core";
 import { LoadingSpinnerModule } from "../../../components/loading-spinner/loading-spinner.module";
 import { SimpleModalModule } from "ngx-simple-modal";
+import { PuzzleComponentsModule } from "../components/puzzle-components.module";
 import { InfoPuzzleHeroModule } from "./components/info-puzzle-hero/info-puzzle-hero.module";
 
 @NgModule({
@@ -31,14 +29,15 @@ import { InfoPuzzleHeroModule } from "./components/info-puzzle-hero/info-puzzle-
         FlexLayoutModule,
         LoadingSpinnerModule,
         TranslateModule,
-        SimpleModalModule,
+        PuzzleComponentsModule,
         InfoPuzzleHeroModule,
+        SimpleModalModule,
 
         StoreModule.forFeature("puzzleHeroTracks", fromPuzzleHero.reducer),
         EffectsModule.forFeature([TracksEffects])
     ],
-    exports: [],
-    declarations: [TracksComponent, TrackComponent, TrackDirective, PuzzleTileComponent],
+    exports: [TracksComponent],
+    declarations: [TracksComponent],
     providers: []
 })
 export class TracksModule {
