@@ -1,10 +1,6 @@
 import { Component, EventEmitter, Input, OnInit, Output } from "@angular/core";
 import { Activity } from "../../../api/models/activity";
-import { select, Store } from "@ngrx/store";
-import { State } from "../store/schedule.reducer";
-import { getCurrentAttendee } from "../../../store/app.reducers";
-import { Observable } from "rxjs";
-import { map } from "rxjs/operators";
+import { TranslateService } from "@ngx-translate/core";
 
 @Component({
     selector: "app-activity-card",
@@ -20,21 +16,11 @@ export class ActivityCardComponent implements OnInit {
     @Output()
     public info = new EventEmitter();
 
-    public currentAttendee$ = this.store$.pipe(select(getCurrentAttendee));
+    constructor(private translateService: TranslateService) { }
 
-    public get subscribed(): Observable<boolean> {
-        return this.currentAttendee$.pipe(map(attendee => {
-            if (!attendee) {
-                return false;
-            }
+    public ngOnInit() { }
 
-            return this.activity.subscribers.some(x => x === attendee._id);
-        }));
-    }
-
-    constructor(private store$: Store<State>) {
-    }
-
-    public ngOnInit() {
+    public get lang(): string {
+        return this.translateService.getDefaultLang();
     }
 }
