@@ -2,9 +2,15 @@ import { Injectable } from "@angular/core";
 import { Observable } from "rxjs";
 import { map } from "rxjs/operators";
 import { ApiService } from "../api/api.service";
-import { PuzzleHeroInfo, Score, TeamSeries, Track } from "../api/models/puzzle-hero";
+import { PuzzleHero, PuzzleHeroInfo, Score, TeamSeries, Track, PuzzleInfo } from "../api/models/puzzle-hero";
 import * as io from "socket.io-client";
 import { environment } from "../../environments/environment";
+import { PuzzleHeroUtils } from "../features/puzzle-hero/utils/puzzle-hero.utils";
+import { TrackFormDto } from "../features/puzzle-hero/admin/components/track-form/dto/track-form.dto";
+import { CreateTrackDto, UpdateTrackDto, CreatePuzzleDto } from "../api/dto/puzzle-hero";
+import { PuzzleAdminUtils } from "../features/puzzle-hero/admin/puzzle-admin.utils";
+import { PuzzleHeroSettingsDto } from "../features/puzzle-hero/admin/components/puzzle-hero-settings/dto/puzzle-hero-settings.dto";
+import { QuestionFormDto } from "../components/question-form/dto/question-form.dto";
 
 const STARRED_TRACKS = "STARRED_TRACKS";
 
@@ -35,20 +41,13 @@ export class PuzzleHeroService {
         return this.apiService.puzzleHero.getInfo();
     }
 
+    getPuzzleHero(): Observable<PuzzleHero> {
+        return this.apiService.puzzleHero.getPuzzleHero();
+    }
+
     getTracks(): Observable<Track[]> {
         return this.apiService.puzzleHero.getPuzzleHero().pipe(
-            map(x => x.tracks),
-            map(x => x.map(track => {
-                return {
-                    ...track,
-                    puzzles: track.puzzles.map(puzzle => {
-                        return {
-                            ...puzzle,
-                            id: (puzzle as any)._id
-                        };
-                    })
-                };
-            }))
+            map(puzzleHero => PuzzleHeroUtils.formatPuzzleHeroTracksIds(puzzleHero))
         );
     }
 
@@ -104,4 +103,28 @@ export class PuzzleHeroService {
     validatePuzzleHero(puzzleId: string, answer: string): Observable<void> {
         return this.apiService.puzzleHero.validatePuzzleHero(puzzleId, answer);
     }
+
+    createTrack(trackFormDto: TrackFormDto): Observable<Track> {
+        return this.apiService.puzzleHero.createTrack(PuzzleAdminUtils.trackFormDtoToTrackDto(trackFormDto) as CreateTrackDto);
+    }
+
+    updateTrack(id: string, trackFormDto: TrackFormDto): Observable<void> {
+        return this.apiService.puzzleHero.updateTrack(id, PuzzleAdminUtils.trackFormDtoToTrackDto(trackFormDto) as UpdateTrackDto);
+    }
+
+    updatePuzzleHero(puzzleHeroSettingsDto: PuzzleHeroSettingsDto): Observable<void> {
+        return this.apiService.puzzleHero.updatePuzzleHero(PuzzleAdminUtils.puzzleHeroSettingsDtoToPuzzleHero(puzzleHeroSettingsDto));
+    }
+
+    createPuzzle(trackId: string, parentId: string, questionFormDto: QuestionFormDto): Observable<PuzzleInfo> {
+        return this.apiService.puzzleHero.createPuzzle(
+            trackId,
+            PuzzleAdminUtils.puzzleFormDtoToPuzzleDto(parentId, questionFormDto) as CreatePuzzleDto
+        );
+    }
+
+    updatePuzzle(trackId: string, id: string, questionFormDto: QuestionFormDto): Observable<void> {
+        return this.apiService.puzzleHero.updatePuzzle(trackId, id, PuzzleAdminUtils.puzzleFormDtoToUpdateQuestionDto(questionFormDto));
+    }
+
 }
