@@ -13,6 +13,7 @@ import { Event } from "../api/models/event";
 import { PuzzleHeroInfo } from "../api/models/puzzle-hero";
 import { ChangePasswordComponent } from "../features/dashboard/modals/change-password/change-password.component";
 import { ProfileSettingComponent } from "../features/dashboard/modals/profile-setting/profile-setting.component";
+import { TicketComponent } from "../features/dashboard/modals/ticket/ticket.component";
 import { AttendeeService } from "../providers/attendee.service";
 import { AuthenticationService } from "../providers/authentication.service";
 import { EventService } from "../providers/event.service";
@@ -21,9 +22,12 @@ import { PuzzleHeroService } from "../providers/puzzle-hero.service";
 import {
     AllNotificationsSeen, AppActionTypes, AppLoaded, ChangeLanguage, ChangePassword, CheckUnseenNotification, CurrentAttendeeLoaded,
     EditProfile, EventsLoaded, GetPuzzleHeroInfo, GlobalError, HasUnseenNotification, InitializeMessaging, LoadCurrentAttendee, LoadEvents,
-    Logout, SetCurrentEvent, SetupMessagingToken, UpdatePuzzleHeroStatus
+    Logout, SetCurrentEvent, SetupMessagingToken, UpdatePuzzleHeroStatus, LoadRegisteredCompetitions, RegisteredCompetitionsLoaded,
+    ViewTicket
 } from "./app.actions";
 import { getCurrentAttendee, getEvents, getPuzzleHeroInfo, State } from "./app.reducers";
+import { CompetitionsService } from "../providers/competitions.service";
+import { Competition } from "../api/models/competition";
 
 @Injectable()
 export class AppEffects {
@@ -65,6 +69,7 @@ export class AppEffects {
                 await this.attendeeService.removeMessagingToken(token).toPromise();
             }
         }),
+        delay(500),
         switchMap(() => {
             return this.authenticationService.logout().pipe(
                 tap(() => {
@@ -150,7 +155,7 @@ export class AppEffects {
     setCurrentEvent$ = this.actions$.pipe(
         ofType<SetCurrentEvent>(AppActionTypes.SetCurrentEvent),
         tap((action) => this.eventService.saveCurrentEvent(action.event._id)),
-        switchMap(() => [new LoadCurrentAttendee(), new GetPuzzleHeroInfo()])
+        switchMap(() => [new LoadCurrentAttendee(), new GetPuzzleHeroInfo(), new LoadRegisteredCompetitions()])
     );
 
     @Effect({ dispatch: false })
@@ -166,6 +171,14 @@ export class AppEffects {
         ofType<ChangePassword>(AppActionTypes.ChangePassword),
         map(() => {
             this.modalService.addModal(ChangePasswordComponent);
+        })
+    );
+
+    @Effect({ dispatch: false })
+    viewTicket$ = this.actions$.pipe(
+        ofType<ViewTicket>(AppActionTypes.ViewTicket),
+        map(() => {
+            this.modalService.addModal(TicketComponent);
         })
     );
 
@@ -198,6 +211,16 @@ export class AppEffects {
         switchMap(() => {
             return this.puzzleHeroService.getInfo().pipe(
                 map((info: PuzzleHeroInfo) => new UpdatePuzzleHeroStatus(info))
+            );
+        }),
+    );
+
+    @Effect()
+    loadRegisteredCompetitions$ = this.actions$.pipe(
+        ofType<LoadRegisteredCompetitions>(AppActionTypes.LoadRegisteredCompetitions),
+        switchMap(() => {
+            return this.eventService.getRegisteredCompetitions().pipe(
+                map((competitions: Competition[]) => new RegisteredCompetitionsLoaded(competitions))
             );
         }),
     );
