@@ -4,6 +4,7 @@ import { Attendee } from "../api/models/attendee";
 import { Event } from "../api/models/event";
 import { PuzzleHeroInfo } from "../api/models/puzzle-hero";
 import { AppActions, AppActionTypes } from "./app.actions";
+import { Competition } from "../api/models/competition";
 
 export interface GlobalState {
     currentAttendee: Attendee;
@@ -13,6 +14,7 @@ export interface GlobalState {
     language: string;
     unseen: boolean;
     puzzleHeroInfo: PuzzleHeroInfo;
+    registeredCompetitions: Competition[];
 }
 
 export interface State {
@@ -30,7 +32,8 @@ export const initialState: GlobalState = {
     loading: true,
     language: null,
     unseen: false,
-    puzzleHeroInfo: null
+    puzzleHeroInfo: null,
+    registeredCompetitions: []
 };
 
 export function globalReducer(state = initialState, action: AppActions): GlobalState {
@@ -90,6 +93,17 @@ export function globalReducer(state = initialState, action: AppActions): GlobalS
                 ...state,
                 puzzleHeroInfo: action.payload
             };
+        case AppActionTypes.LoadRegisteredCompetitions:
+            return {
+                ...state,
+                loading: true
+            };
+        case AppActionTypes.RegisteredCompetitionsLoaded:
+            return {
+                ...state,
+                loading: false,
+                registeredCompetitions: action.payload
+            };
     }
     return state;
 }
@@ -116,3 +130,5 @@ export const getCurrentLanguage = createSelector(getGlobalState, (state: GlobalS
 export const getUnseen = createSelector(getGlobalState, (state: GlobalState) => state.unseen);
 
 export const getPuzzleHeroInfo = createSelector(getGlobalState, (state: GlobalState) => state.puzzleHeroInfo);
+
+export const getRegisteredCompetitions = createSelector(getGlobalState, (state: GlobalState) => state.registeredCompetitions);
